add fixtures
