+++ tests/test_processing.py
@@ -1,18 +1,5 @@
-import pytest
-
 from src.processing import filter_by_state, sort_by_date
-
-
-@pytest.fixture()
-def list_dicts():
-    return [
-        {},
-        {"id": 41428829},
-        {"state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"state": "EXECUTED"},
-        {"state": "CANCELED"},
-        {"state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
-    ]
+from .fixtures import list_dicts_date, list_dicts
 
 
 def test_filter_by_state_default_state(list_dicts):
@@ -30,18 +17,8 @@ def test_sort_by_date_is_empty():
     assert sort_by_date([]) == []
 
 
-@pytest.fixture
-def test_list_dicts_date():
-    return [
-        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
-        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-    ]
-
-
-def test_sort_by_date_key_date(test_list_dicts_date, reverse=True):
-    assert sort_by_date(test_list_dicts_date, ascending=True) == [
+def test_sort_by_date_key_date(list_dicts_date, reverse=True):
+    assert sort_by_date(list_dicts_date, ascending=True) == [
         {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
         {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
         {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
